+++ src/capstone2llvmir/arm/arm_impl.h
@@ -161,6 +161,7 @@ class Capstone2LlvmIrTranslatorArm_impl :
 		void translateSbc(cs_insn* i, cs_arm* ai, llvm::IRBuilder<>& irb);
 		void translateShifts(cs_insn* i, cs_arm* ai, llvm::IRBuilder<>& irb);
 		void translateStr(cs_insn* i, cs_arm* ai, llvm::IRBuilder<>& irb);
+		void translateStrex(cs_insn* i, cs_arm* ai, llvm::IRBuilder<>& irb);
 		void translateSub(cs_insn* i, cs_arm* ai, llvm::IRBuilder<>& irb);
 		void translateUmlal(cs_insn* i, cs_arm* ai, llvm::IRBuilder<>& irb);
 		void translateUmull(cs_insn* i, cs_arm* ai, llvm::IRBuilder<>& irb);